Move events out of the public api and into the lifecycle.
