+++ test/unit.js
@@ -1,7 +1,6 @@
 import './boot';
 import './unit/api/create';
 import './unit/api/emit';
-import './unit/api/event';
 import './unit/api/init';
 import './unit/api/no-conflict';
 import './unit/api/property';
@@ -12,6 +11,7 @@ import './unit/dom';
 import './unit/extending';
 import './unit/ignoring';
 import './unit/lifecycle';
+import './unit/lifecycle/events';
 import './unit/registration';
 import './unit/registry';
 import './unit/templating';
